- Refactor command scopes
